+++ client/client_test.go
@@ -1,29 +1,30 @@
-package client
+package client_test
 
 import (
 	"reflect"
 	"testing"
 
-	"github.com/m1/go-finnhub/crypto"
-	"github.com/m1/go-finnhub/forex"
-	"github.com/m1/go-finnhub/news"
-	"github.com/m1/go-finnhub/stock"
+	"github.com/llonchj/go-finnhub/client"
+	"github.com/llonchj/go-finnhub/crypto"
+	"github.com/llonchj/go-finnhub/forex"
+	"github.com/llonchj/go-finnhub/news"
+	"github.com/llonchj/go-finnhub/stock"
 )
 
 func TestNew(t *testing.T) {
-	a := NewAPI("token", Version)
+	a := client.NewAPI("token", client.Version)
 	type args struct {
 		key string
 	}
 	tests := []struct {
 		name string
 		args args
-		want *Client
+		want *client.Client
 	}{
 		{
 			name: "valid",
 			args: args{key: "token"},
-			want: &Client{
+			want: &client.Client{
 				Stock:  stock.Client{API: a},
 				Forex:  forex.Client{API: a},
 				Crypto: crypto.Client{API: a},
@@ -33,7 +34,7 @@ func TestNew(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if got := New(tt.args.key); !reflect.DeepEqual(got, tt.want) {
+			if got := client.New(tt.args.key); !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("New() = %v, want %v", got, tt.want)
 			}
 		})