+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 # Changelog
 
+## [0.3.0] - 2019-04-13
+
+- Moved to github.com/llonchj/go-finnhub
+- Updated docs.
+- Access to http client & simplified tests
+- Deprecated GetCEO in favour of GetExecutive
+
+[0.3.0]: https://github.com/llonchj/go-finnhub/tree/0.3.0
+
 ## [0.2.0] - 2019-01-14
+
 - Updated tests and docs.
 
 [0.2.0]: https://github.com/m1/go-finnhub/tree/0.2.0
 
 ## [0.1.0] - 2019-01-13
+
 - Initial release
 
 [0.1.0]: https://github.com/m1/go-finnhub/tree/0.1.0